Refactor in upgrading scripts
